Revert "some other improvements on the backend docker setup"
This reverts commit d32bba14641ed31d9b35ef51dddb1ec627b6e971.

+++ docker/backend/backend.sh
@@ -12,7 +12,10 @@ if [ ! -d "venv" ]; then
 fi
 
 source venv/bin/activate
-pip install -r requirements.txt
+
+if ! pip freeze | grep -q -F -r requirements.txt; then
+    pip install -r requirements.txt
+fi
 
 echo 'source /app/backend/venv/bin/activate' >> /root/.bashrc
